refact(views/marca):correção em deletar marca  #77

+++ sgo_systems/marca/views.py
@@ -2,7 +2,7 @@ from django.shortcuts import redirect, render
 from .models import Marca
 from .forms import MarcaForm
 from django.contrib import messages
-# Create your views here.
+from django.shortcuts import get_object_or_404
 
 def cadastrar_marca(request):
   if request.method == "POST":
@@ -20,9 +20,14 @@ def listar_marcas(request):
   return render(request, "listar_marcas.html", {"marcas": marcas})
 
 def deletar_marca(request, pk):
-  marca = Marca.objects.get(pk=pk)
+  marca = get_object_or_404(Marca, pk=pk)
+  if marca.produtos_set.exists():  
+    messages.error(request, "Não é possível excluir. Há produtos vinculados a esta marca.")
+    return redirect("listar_marcas")
   marca.delete()
+  messages.success(request, "Marca excluída com sucesso.")
   return redirect("listar_marcas")
+
 
 def editar_marca(request, pk):
     marca = marca.objects.get(pk=pk)
